Add a PCAP to show the use of X-Forwarded-For
X-Forwarded-For is used to inform the target server of the IP address of
the requesting client. In this PCAP, the requests for the '/' path
contain this header. Requests for /index.html do not contain this header
as those requests originate from the load balancer itself when it
performs health checks, and are therefore not forarded for any client.